+++ src/lib.rs
@@ -32,7 +32,4 @@ pub mod mqtt;
 pub mod plot;
 #[cfg(all(feature = "profiling", not(target_arch = "wasm32")))]
 pub mod profiling;
-#[cfg(feature = "selfupdater")]
-#[cfg(not(target_arch = "wasm32"))]
-pub mod updater;
 pub mod util;